wx/requirements.txt
loguru
Appium-Python-Client
tqdm
parse
